Fazendo as atualizações necessarios, pois agora é LONG não mais Integer

+++ src/main/java/com/br/SAM_FullStack/SAM_FullStack/repository/AreaDeAtuacaoRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AreaDeAtuacaoRepository extends JpaRepository<AreaDeAtuacao, Integer> {
+public interface AreaDeAtuacaoRepository extends JpaRepository<AreaDeAtuacao, Long> {
 }
